+++ chromeos/ash/components/quick_start/quick_start_requests.h
@@ -22,7 +22,7 @@ std::unique_ptr<QuickStartMessage> BuildAssertionRequestMessage(
 std::unique_ptr<QuickStartMessage> BuildGetInfoRequestMessage();
 
 std::unique_ptr<QuickStartMessage> BuildRequestWifiCredentialsMessage(
-    int32_t session_id,
+    uint64_t session_id,
     std::string& shared_secret);
 
 std::vector<uint8_t> CBOREncodeGetAssertionRequest(const cbor::Value& request);
@@ -31,7 +31,7 @@ cbor::Value GenerateGetAssertionRequest(
     std::array<uint8_t, crypto::kSHA256Length> client_data_hash);
 
 std::unique_ptr<QuickStartMessage> BuildNotifySourceOfUpdateMessage(
-    int32_t session_id,
+    uint64_t session_id,
     const base::span<uint8_t, 32> shared_secret);
 }  // namespace ash::quick_start::requests
